refactor: padronizar o tipo das chaves primárias como Long
Os ids estavam divididos entre Integer e Long.

- Troca Integer por Long em Estoque, Fornecedor, Funcionario, Tinta
  e Usuario
- Ajusta o segundo parâmetro dos JpaRepository e as assinaturas de
  findById e inativar nos services e controllers

+++ src/main/java/com/verf_system/verfS/controller/EstoqueController.java
@@ -24,7 +24,7 @@ public class EstoqueController {
 
     @GetMapping(value = "ID/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public EstoqueEntity findById(@PathVariable Integer id) {
+    public EstoqueEntity findById(@PathVariable Long id) {
         return estoqueService.findById(id);
     }
