+++ src/main/java/com/verf_system/verfS/database/repository/IUsuarioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IUsuarioRepository extends JpaRepository<UsuarioEntity, Integer> {
+public interface IUsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
 }
